Remove Angular static files, code cleanup
