--- canvas_site_creator/static/canvas_site_creator/js/controllers/ErrorController.js
@@ -1,15 +1,0 @@
-(function(){
-    /**
-     * Angular controller for displaying errors.
-     */
-    angular.module('app').controller('ErrorController', ['$scope', 'errorModel', function($scope, errorModel){
-        $scope.errorModel = errorModel;
-
-        // Output DataTable AJAX errors to the browser console instead of as Javascript alerts
-        $.fn.dataTable.ext.errMode = function(settings, helpPage, message){
-            console.log(message);
-            $scope.errorModel.hasError = true;
-            $scope.$apply();
-        };
-    }]);
-})();
