--- canvas_site_creator/static/canvas_site_creator/js/models/BulkJobDetailModel.js
@@ -1,14 +1,0 @@
-(function(){
-    /**
-     * Angular service for sharing data needed for rendering and controlling bulk job detail data across controllers.
-     */
-    angular.module('app').factory('bulkJobDetailModel', [function(){
-        return {
-            totalCourseJobs: 0,
-            completeCourseJobs: 0,
-            successfulCourseJobs: 0,
-            failedCourseJobs: 0,
-            percentageComplete: 0
-        };
-    }]);
-})();
