--- canvas_site_creator/static/canvas_site_creator/js/models/ErrorModel.js
@@ -1,10 +1,0 @@
-(function(){
-    /**
-     * Angular service for sharing data needed for rendering and controlling errors across controllers.
-     */
-    angular.module('app').factory('errorModel', [function(){
-        return {
-            hasError: false
-        };
-    }]);
-})();
